+++ app/src/main/res/layout/activity_main.xml
@@ -4,4 +4,4 @@
     android:id="@+id/rootLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.MainActivity" />
+    tools:context="activities.MainActivity" />
